+++ lib/src/test/test.config.dart
@@ -1,0 +1,7 @@
+class TestConfig {
+  static const String emailA = 'apple@email.com';
+  static const String emailB = 'banana@email.com';
+  static const String emailC = 'cherry@email.com';
+  static const String emailD = 'durian@email.com';
+  static const String password = 'p/Dux,*3u.:Et';
+}